# swagger-gateway.yaml
swagger: "2.0"
info:
  title: Gateway API
  description: Piazza Core Services API
  version: "0.1.0"
  contact: 
    name: Patrick Doody
    email: patrick.doody@rgi-corp.com
    url: https://github.com/venicegeo/venice/wiki
host: pz-gateway.stage.geointservices.io
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /job/{jobId}:
    get:
      summary: Get Job Status
      description: "Fetches the Status for a single Piazza Job."
      parameters:
        - name: jobId
          in: path
          description: "ID of the Job to Fetch"
          required: true
          type: string
      tags:
        - Job
      responses:
        200:
          description: "Information regarding the requested Job. At bare minimum, this will contain the Job ID of the Job that has been spawned as a result of the POSTed message. If more information is available, such as Status, it will also be included."
          schema:
            $ref: '#/definitions/JobStatus'
        default:
          description: "Unexpected Error"
          schema:
            $ref: '#/definitions/ErrorResponse'
    put:
      summary: Repeat Job
      description: "Repeats a previously submitted Job. This will clone the original Job, and run it again with identical parameters, using the requesting users authentication in the new Job."
      parameters:
        - name: jobId
          in: path
          description: "ID of the Job to Repeat"
          required: true
          type: string
      tags:
        - Job
      responses:
        200:
          description: "A new Job ID that corresponds to the cloned Job in Piazza."
          schema:
            $ref: '#/definitions/JobResponse'
        default:
          description: "Unexpected Error."
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      summary: Abort Job
      description: "Cancels a Running Job. If the Job is already completed in some way, then cancellation will not occur."
      parameters:
        - name: jobId
          in: path
          description: "ID of the Job to cancel."
          required: true
          type: string
        - name: reason
          in: query
          description: "Details for the cancellation of the Job."
          required: false
          type: string
      tags:
        - Job
      responses:
        200:
          description: "The Job has requested to be cancelled. This may take some time, as the process may not be in an easily cancelled state at the time the request is made."
        default:
          description: "Unexpected Error"
          schema:
            $ref: '#/definitions/ErrorResponse'
  /v2/job:
    post:
      summary: Executes a registered Service
      description: "Creates a Piazza Job to execute a registered service in the system, with the specified parameters."
      parameters:
        - name: data
          in: body
          description: "The Payload that describes the Service to be executed, and the inputs for that service."
          required: true
          schema:
            $ref: '#/definitions/ExecuteJob'
      tags:
        - Job
        - Service
      responses:
        200:
          description: "The Job ID for the execution of the Service. This can be queried using Job Status to track progress and, when available, fetch the result object."
          schema:
            $ref: '#/definitions/JobResponse'
        default:
          description: "Unexpected Error"
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
  ErrorResponse:
    description: "The standard response object for an unexpected error within the Piazza core."
    type: object
    required:
      - message
      - origin
    properties:
      message:
        description: "The high-level description of the error message. May contain debug information useful for technical debugging."
        type: string
      origin:
        description: "The name of the component the error originated from."
        type: string
  JobResponse:
    type: object
    description: "An asynchronous response from the Gateway where a Job has been created, with the provided Job ID. This Job ID can then be queried for Job status to determine when the Job has succeeded, or to track its progress."
    required:
      - jobId
    properties:
      jobId:
        description: "The ID of the Job that has been created from the request."
        type: string
...
  JobStatus:
    type: object
    description: "Represents the status of a Job in the Piazza system; either running or completed. Contains information such as progress, time remaining, errors encountered, Job results, or general Job Status."
    required:
      - status
      - jobType
      - submittedBy
    properties:
      result:
        description: "A reference to the Result of the Job. This could be a Resource ID, or a Service ID, in certain cases. Or perhaps an error if the Job encountered an error during processing."
        type: object
      status:
        description: "The status of the Job. Submitted, Running, Success, Error, or Failure."
        type: string
      jobType:
        description: "Polymorphically describes the information defining what actions this Job should take."
        type: string
      submittedBy:
        description: "The name of the user who submitted the Job."
        type: string
      progress:
        $ref: '#/definitions/JobProgress'
  JobProgress:
    type: object
    description: "Defines Job Progress."
    properties:
      percentComplete:
        description: "The percentage complete for the Job, in whole numbers, out of 100 percent"
        type: integer
        format: int32
      timeRemaining:
        description: "The last value for the Time Remaining as reported by the processor of this Job."
        type: string
      timeSpent:
        description: "The amount of time spent by the processor of this Job, according to that processor."
        type: string
...
  ExecuteJob:
    type: object
    description: "The payload that describes a Service Controller execution Job."
    properties:
      serviceId:
        type: string
        description: "The ID of the Service to Execute."
      dataInputs:
        type: object
        description: "The Inputs for the execution."
      dataOutput: 
        type: object
        description: "The Output for the execution."
...